创建课件接口
