+++ src/config.js
@@ -1,1 +1,1 @@
-export var localhost = "192.168.1.114"
+export var localhost = "192.168.32.199"
