+++ src/pages/CreatecourseNew/pagefirst.js
@@ -55,136 +55,6 @@ class pagefirst extends Component {
     this.state.step_content.push(step2)
     this.state.step_content.push(step3)
   }
-
-  createcourse = () =>{
-    const know=this.state.knowledgelist;
-    var knowsource=[];
-    for(var i=0;i<know.length;i++){
-       var obj=know[i];
-       knowsource.push(obj.title);
-    };
-    console.log(knowsource)
-    const { login_info }=this.props;
-    var data={
-     "user_id":login_info.user_id,
-      "courseName":this.state.courseName,
-      "grade": this.state.grade,
-      "subject": this.state.subject,
-      "descript":  this.state.descript,
-      "knowledges":knowsource,
-      "isOpen": this.state.isOpen,
-      "isEdit": 1,
-      "name": "课件目录",
-      "children":this.state.coursecatalog,
-      "templateId": 1,
-      "slide": [{
-          "pageId": 1,
-          "pageThumbnail": {
-              "pageurl": "./1.png",
-              "style": {
-                  "pagewidth": "100px",
-                  "pageheight": "100px"
-              }
-          },
-          "media":[
-              {
-                  "id":2314,
-                  "position":[0,0],
-                  "rotation":0,
-                  "scale":[1,1],
-                  "shape":{"cx":100,"cy":100,"n":30,"z":40},
-                  "style":{"fill":"none"},
-                  "type":"house"
-              }
-          ]
-      }],     
-      "fileSize": "100M",
-      "scope": "k12教育",
-      "addTime": new Date(),
-      "views": 300,
-      "url": "D:/Graduate/11.jpg",
-      "width": "30px",
-      "height": "40px"
-  };
-   
-    //创建课件
-    console.log("进入ajax") 
-    const {setCreatecourseState} = this.props;
-    $.ajax({
-        url: "http://"+localhost+":3000/api/createCourse",
-        async:false,
-        type: "POST",
-        contentType:"application/json;charset=UTF-8",
-        accepts:"application/json;charset=UTF-8",
-        dataType: "json",
-        data:JSON.stringify(data),
-        beforeSend:function(request){
-          request.setRequestHeader("Authorization",'Bearer '+login_info.access_token);
-        },
-        success: function (data) {
-            if (data.errorCode === 0) {
-                console.log('成功保存课件');
-                console.log(data.msg);
-                console.log(data.msg._id);
-                message.success('成功创建课件！');
-               
-                setCreatecourseState({
-                  type:'createcourseSuccess',
-                  payload:{
-                    createCourse_info:data.msg,
-                    course_id:data.msg._id,
-                    // numchat:false,
-                  }
-                });
-                this.context.router.history.push("/APP");
-
-            }
-            else {
-                console.log('成功获取搜索资源');
-                // this.setState({ resource: data.msg });
-                console.log(data.msg);
-            }
-        }.bind(this),
-        error: function (xhr, status, err) {
-          console.log("取回课件数据错误")
-        }
-    });
-}
-
-getknowledgeRel(value) {
-  const { login_info }=this.props;
-  console.log('进入knowledgeRel ajax');
-  console.log(login_info.access_token);
-  $.ajax({
-    url: "http://"+localhost+":3000/api/knowledgeRel",
-    data:"courseName="+value,
-    beforeSend:function(request){
-      request.setRequestHeader("Authorization",'Bearer '+login_info.access_token);
-    },
-    type: "GET",
-    dataType: "json",
-    async:false,
-    success: function (data) {
-      if (data.errorCode === 0) {
-        console.log('获取关联知识点');
-        console.log(data);
-        if(data.msg){
-        this.setState({
-          knowledgelist:data.msg,
-        });
-      }
-      }
-      else {   
-        console.log('获取关联知识点2222');
-      }
-    }.bind(this),
-    error: function (xhr, status, err) {
-    }
-  });
-}
-
-
-
   render() {
     const { current } = this.state;
     return (
@@ -215,12 +85,10 @@ const pagefirst_Index=Form.create()(pagefirst);
   function  mapStateToProps(state) {
     return{
        login_info:state.reducer_login.login_info,
-     //  user_info:state.reducer_user.user_info,
     };
   }
   function mapDispatchToProps(dispatch){
     return{
-      setCreatecourseState: (state) => dispatch(state),
     };
   }
   export default connect(
